Removed string methods not covered from comments Added trim Fixed integer division issue in switch

+++ Week01/W01_1/W_01_1_03_Switch.cs
@@ -4,9 +4,9 @@ public static class W_01_1_03_Switch
     {
         // What does the following code do?
         Console.WriteLine("Enter a number");
-        int num1 = Convert.ToInt32(Console.ReadLine());
+        double num1 = Convert.ToDouble(Console.ReadLine());
         Console.WriteLine("Enter another number");
-        int num2 = Convert.ToInt32(Console.ReadLine());
+        double num2 = Convert.ToDouble(Console.ReadLine());
 
         Console.WriteLine("Enter a symbol");
         string symbol = Console.ReadLine();
